+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Layer context menu action to update its first/last frame based on the main comp
 * I/O:
     * Animated SVG now honours layer first/last frames
+    * SVG keeps track of parent layer transforms
 * Scripting:
     * Python has now access to the users of an asset
 * Bug Fixes: